cpp-port: batch 69 - port 6 hw/ files to C++
Files ported:
- hw/misc/aspeed_sli.cpp
- hw/nubus/nubus-bus.cpp
- hw/xen/xen-bus-helper.cpp
- hw/usb/combined-packet.cpp
- hw/isa/isa-superio.cpp
- hw/hyperv/hv-balloon-our_range_memslots.cpp

10 headers with extern "C" guards added.
All files include trace.h (previously blocked, now resolved).

+++ hw/usb/combined-packet.cpp
@@ -68,7 +68,7 @@ void usb_combined_input_packet_complete(USBDevice *dev, USBPacket *p)
     QTAILQ_FOREACH_SAFE(p, &combined->packets, combined_entry, next) {
         if (!done) {
             /* Distribute data over uncombined packets */
-            if (actual_length >= p->iov.size) {
+            if (static_cast<size_t>(actual_length) >= p->iov.size) {
                 p->actual_length = p->iov.size;
             } else {
                 /* Send short or error packet to complete the transfer */
